+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.4.0] - 2018-04-27
+
+### Added
+
+* `RegularExpression` scalar type _generator_
+* `UnsignedInt` alias for `NonNegativeInt`
+* `UnsignedFloat` alias for `NonNegativeFloat`
+
 ## [0.3.0] - 2018-04-06
 
 ### Changed